django-testing/django_testing/requirements-dev.txt
-r requirements.txt
pytest
pytest-django
model_bakery
pytest-cov